First version of connectivity-manager
Goal with this first version is to support:

- Enable/disable Wi-Fi.
- Listing Wi-Fi access points.
- Connect to an access point and prompt for password if required.
- Disconnect from an access point.

The D-Bus API required to accomplish this is added and implemented in the
service and a simple command line interface is included for testing it.

A backend abstraction is added in the daemon to allow for changing
"network manager" implementation. The first concrete implementation is a
backend for ConnMan since it is what we currently use in Pelux. The
thinking is that it should be possible to have e.g. a NetworkManager
implementation or some other custom/vendor specific implementation in the
future.

There are improvements to be made in the ConnMan backend and a lot still
needs to be tested. Have added TODO:s in the code for some of it.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>