fixup! First version of connectivity-manager
Fix copy paste error that meant default_value was not returned when looking
up ConnMan agent field arguments and it had the wrong type (not a string).
Currently always pass "" as the default value so there is no difference in
practice at the moment to change from {} to default_value. But if another
default value is ever passed in the future, it would be wrong.