Conditionally include header for ConnManBackend in src/daemon/backend.cpp
It does not really matter for the ConnMan backend since it e.g. does not
use any additional library whose headers are included in connman_backend.h.
But if such a backend were to be added, it would cause a compilation error
when compiling with some other backend since the dependency would not be
available. For instance, NetworkManager has libnm. Do not want to
complicate things with PIMPL or similar patterns when there is no need.

Signed-off-by: Martin Ejdestig <marejde@gmail.com>